requirements.txt
flask==3.0.0
python-dotenv==1.0.0
groq>=0.3.0
pytesseract==0.3.10
Pillow==10.0.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
